Improve examples
