+++ Example/src/Panel4.tsx
@@ -2,23 +2,14 @@ import React, { useState } from 'react';
 import { Modal, Pressable, StyleSheet, Text, View } from 'react-native';
 import Animated, { useAnimatedStyle, useSharedValue } from 'react-native-reanimated';
 
-import ColorPicker, {
-  Swatches,
-  OpacitySlider,
-  colorKit,
-  PreviewText,
-  RedSlider,
-  GreenSlider,
-  BlueSlider,
-} from 'reanimated-color-picker';
+import ColorPicker, { Panel4, OpacitySlider, colorKit, PreviewText } from 'reanimated-color-picker';
 import type { returnedResults } from 'reanimated-color-picker';
 
 export default function Example() {
   const [showModal, setShowModal] = useState(false);
 
-  const customSwatches = new Array(6).fill('#fff').map(() => colorKit.randomRgbColor().hex());
-
-  const selectedColor = useSharedValue(customSwatches[0]);
+  const initialColor = colorKit.randomHsvColor({ s: [100, 100], v: [100, 100] }).hex();
+  const selectedColor = useSharedValue(initialColor);
   const backgroundColorStyle = useAnimatedStyle(() => ({ backgroundColor: selectedColor.value }));
 
   const onColorSelect = (color: returnedResults) => {
@@ -29,7 +20,7 @@ export default function Example() {
   return (
     <>
       <Pressable style={styles.openButton} onPress={() => setShowModal(true)}>
-        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Horizontal RGB</Text>
+        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Panel4</Text>
       </Pressable>
 
       <Modal onRequestClose={() => setShowModal(false)} visible={showModal} animationType='slide'>
@@ -41,25 +32,13 @@ export default function Example() {
               thumbSize={24}
               thumbShape='circle'
               onChange={onColorSelect}
-              thumbAnimationDuration={100}
-              adaptSpectrum
-              boundedThumb
             >
-              <Text style={styles.sliderTitle}>Red</Text>
-              <RedSlider style={styles.sliderStyle} />
+              <Panel4 style={styles.panelStyle} thumbShape='ring' />
 
-              <Text style={styles.sliderTitle}>Green</Text>
-              <GreenSlider style={styles.sliderStyle} />
-
-              <Text style={styles.sliderTitle}>Blue</Text>
-              <BlueSlider style={styles.sliderStyle} />
-
-              <Text style={styles.sliderTitle}>Opacity</Text>
               <OpacitySlider style={styles.sliderStyle} />
 
-              <Swatches style={styles.swatchesContainer} swatchStyle={styles.swatchStyle} colors={customSwatches} />
               <View style={styles.previewTxtContainer}>
-                <PreviewText style={{ color: '#707070' }} colorFormat='rgba' />
+                <PreviewText style={{ color: '#707070' }} colorFormat='hwba' />
               </View>
             </ColorPicker>
           </View>
@@ -77,7 +56,6 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: 'center',
     alignContent: 'center',
-    backgroundColor: 'orange',
   },
   pickerContainer: {
     alignSelf: 'center',
@@ -95,15 +73,22 @@ const styles = StyleSheet.create({
 
     elevation: 10,
   },
-  sliderTitle: {
-    color: '#000',
-    fontWeight: 'bold',
-    marginBottom: 5,
-    paddingHorizontal: 4,
+  panelStyle: {
+    borderRadius: 16,
+
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
   },
   sliderStyle: {
     borderRadius: 20,
-    marginBottom: 20,
+    marginTop: 20,
 
     shadowColor: '#000',
     shadowOffset: {
@@ -120,24 +105,6 @@ const styles = StyleSheet.create({
     marginTop: 20,
     borderTopWidth: 1,
     borderColor: '#bebdbe',
-  },
-  swatchesContainer: {
-    paddingTop: 20,
-    marginTop: 20,
-    borderTopWidth: 1,
-    borderColor: '#bebdbe',
-    alignItems: 'center',
-    flexWrap: 'nowrap',
-    gap: 10,
-  },
-  swatchStyle: {
-    borderRadius: 20,
-    height: 30,
-    width: 30,
-    margin: 0,
-    marginBottom: 0,
-    marginHorizontal: 0,
-    marginVertical: 0,
   },
   openButton: {
     width: '100%',
@@ -158,7 +125,7 @@ const styles = StyleSheet.create({
   },
   closeButton: {
     position: 'absolute',
-    bottom: 20,
+    bottom: 10,
     borderRadius: 20,
     paddingHorizontal: 40,
     paddingVertical: 10,
