+++ ExampleExpo/src/HslHorizontal.tsx
@@ -2,16 +2,16 @@ import React, { useState } from 'react';
 import { Modal, Pressable, StyleSheet, Text, View } from 'react-native';
 import Animated, { useAnimatedStyle, useSharedValue } from 'react-native-reanimated';
 
-import ColorPicker, {
-  Swatches,
-  OpacitySlider,
-  HueSlider,
-  colorKit,
-  PreviewText,
-  SaturationSlider,
-  BrightnessSlider,
-} from 'reanimated-color-picker';
 import type { returnedResults } from 'reanimated-color-picker';
+import ColorPicker, {
+  HSLSaturationSlider,
+  HueSlider,
+  LuminanceSlider,
+  OpacitySlider,
+  PreviewText,
+  Swatches,
+  colorKit,
+} from 'reanimated-color-picker';
 
 export default function Example() {
   const [showModal, setShowModal] = useState(false);
@@ -29,7 +29,7 @@ export default function Example() {
   return (
     <>
       <Pressable style={styles.openButton} onPress={() => setShowModal(true)}>
-        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Vertical HSB Sliders</Text>
+        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>HSL Horizontal</Text>
       </Pressable>
 
       <Modal onRequestClose={() => setShowModal(false)} visible={showModal} animationType='slide'>
@@ -37,23 +37,28 @@ export default function Example() {
           <View style={styles.pickerContainer}>
             <ColorPicker
               value={selectedColor.value}
-              sliderThickness={30}
-              thumbSize={30}
+              sliderThickness={25}
+              thumbSize={24}
               thumbShape='circle'
               onChange={onColorSelect}
               adaptSpectrum
               boundedThumb
             >
-              <View style={{ flexDirection: 'row', justifyContent: 'space-between', height: 300 }}>
-                <HueSlider style={styles.sliderStyle} vertical reverse />
-                <SaturationSlider style={styles.sliderStyle} vertical reverse />
-                <BrightnessSlider style={styles.sliderStyle} vertical reverse />
-                <OpacitySlider style={styles.sliderStyle} vertical reverse />
-              </View>
+              <Text style={styles.sliderTitle}>Hue</Text>
+              <HueSlider style={styles.sliderStyle} />
+
+              <Text style={styles.sliderTitle}>Saturation</Text>
+              <HSLSaturationSlider style={styles.sliderStyle} reverse />
+
+              <Text style={styles.sliderTitle}>Luminance</Text>
+              <LuminanceSlider style={styles.sliderStyle} />
+
+              <Text style={styles.sliderTitle}>Opacity</Text>
+              <OpacitySlider style={styles.sliderStyle} />
 
               <Swatches style={styles.swatchesContainer} swatchStyle={styles.swatchStyle} colors={customSwatches} />
               <View style={styles.previewTxtContainer}>
-                <PreviewText style={{ color: '#707070' }} colorFormat='hsva' />
+                <PreviewText style={{ color: '#707070' }} colorFormat='hsla' />
               </View>
             </ColorPicker>
           </View>
@@ -89,8 +94,15 @@ const styles = StyleSheet.create({
 
     elevation: 10,
   },
+  sliderTitle: {
+    color: '#000',
+    fontWeight: 'bold',
+    marginBottom: 5,
+    paddingHorizontal: 4,
+  },
   sliderStyle: {
     borderRadius: 20,
+    marginBottom: 20,
 
     shadowColor: '#000',
     shadowOffset: {
@@ -145,7 +157,7 @@ const styles = StyleSheet.create({
   },
   closeButton: {
     position: 'absolute',
-    bottom: 20,
+    bottom: 10,
     borderRadius: 20,
     paddingHorizontal: 40,
     paddingVertical: 10,
