+++ ExampleExpo/src/Panel1.tsx
@@ -2,15 +2,8 @@ import React, { useState } from 'react';
 import { Modal, Pressable, StyleSheet, Text, View } from 'react-native';
 import Animated, { useAnimatedStyle, useSharedValue } from 'react-native-reanimated';
 
-import ColorPicker, {
-  Swatches,
-  OpacitySlider,
-  HueSlider,
-  colorKit,
-  PreviewText,
-  SaturationSlider,
-  BrightnessSlider,
-} from 'reanimated-color-picker';
+import ColorPicker, { Panel1, Swatches, OpacitySlider, HueSlider, colorKit, PreviewText } from 'reanimated-color-picker';
+import type { returnedResults } from 'reanimated-color-picker';
 
 export default function Example() {
   const [showModal, setShowModal] = useState(false);
@@ -20,7 +13,7 @@ export default function Example() {
   const selectedColor = useSharedValue(customSwatches[0]);
   const backgroundColorStyle = useAnimatedStyle(() => ({ backgroundColor: selectedColor.value }));
 
-  const onColorSelect = color => {
+  const onColorSelect = (color: returnedResults) => {
     'worklet';
     selectedColor.value = color.hex;
   };
@@ -28,7 +21,7 @@ export default function Example() {
   return (
     <>
       <Pressable style={styles.openButton} onPress={() => setShowModal(true)}>
-        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Horizontal HSB Sliders</Text>
+        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Panel1</Text>
       </Pressable>
 
       <Modal onRequestClose={() => setShowModal(false)} visible={showModal} animationType='slide'>
@@ -40,24 +33,14 @@ export default function Example() {
               thumbSize={24}
               thumbShape='circle'
               onChange={onColorSelect}
-              adaptSpectrum
               boundedThumb
             >
-              <Text style={styles.sliderTitle}>Hue</Text>
+              <Panel1 style={styles.panelStyle} />
               <HueSlider style={styles.sliderStyle} />
-
-              <Text style={styles.sliderTitle}>Saturation</Text>
-              <SaturationSlider style={styles.sliderStyle} reverse />
-
-              <Text style={styles.sliderTitle}>Brightness</Text>
-              <BrightnessSlider style={styles.sliderStyle} />
-
-              <Text style={styles.sliderTitle}>Opacity</Text>
               <OpacitySlider style={styles.sliderStyle} />
-
               <Swatches style={styles.swatchesContainer} swatchStyle={styles.swatchStyle} colors={customSwatches} />
               <View style={styles.previewTxtContainer}>
-                <PreviewText style={{ color: '#707070' }} colorFormat='rgba' />
+                <PreviewText style={{ color: '#707070' }} />
               </View>
             </ColorPicker>
           </View>
@@ -75,7 +58,6 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: 'center',
     alignContent: 'center',
-    backgroundColor: 'orange',
   },
   pickerContainer: {
     alignSelf: 'center',
@@ -93,15 +75,22 @@ const styles = StyleSheet.create({
 
     elevation: 10,
   },
-  sliderTitle: {
-    color: '#000',
-    fontWeight: 'bold',
-    marginBottom: 5,
-    paddingHorizontal: 4,
+  panelStyle: {
+    borderRadius: 16,
+
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
   },
   sliderStyle: {
     borderRadius: 20,
-    marginBottom: 20,
+    marginTop: 20,
 
     shadowColor: '#000',
     shadowOffset: {
@@ -156,7 +145,7 @@ const styles = StyleSheet.create({
   },
   closeButton: {
     position: 'absolute',
-    bottom: 20,
+    bottom: 10,
     borderRadius: 20,
     paddingHorizontal: 40,
     paddingVertical: 10,
