+++ ExampleExpo/src/Panel4.tsx
@@ -2,23 +2,14 @@ import React, { useState } from 'react';
 import { Modal, Pressable, StyleSheet, Text, View } from 'react-native';
 import Animated, { useAnimatedStyle, useSharedValue } from 'react-native-reanimated';
 
-import ColorPicker, {
-  Swatches,
-  OpacitySlider,
-  colorKit,
-  PreviewText,
-  RedSlider,
-  GreenSlider,
-  BlueSlider,
-} from 'reanimated-color-picker';
+import ColorPicker, { Panel4, OpacitySlider, colorKit, PreviewText } from 'reanimated-color-picker';
 import type { returnedResults } from 'reanimated-color-picker';
 
 export default function Example() {
   const [showModal, setShowModal] = useState(false);
 
-  const customSwatches = new Array(6).fill('#fff').map(() => colorKit.randomRgbColor().hex());
-
-  const selectedColor = useSharedValue(customSwatches[0]);
+  const initialColor = colorKit.randomHsvColor({ s: [100, 100], v: [100, 100] }).hex();
+  const selectedColor = useSharedValue(initialColor);
   const backgroundColorStyle = useAnimatedStyle(() => ({ backgroundColor: selectedColor.value }));
 
   const onColorSelect = (color: returnedResults) => {
@@ -29,7 +20,7 @@ export default function Example() {
   return (
     <>
       <Pressable style={styles.openButton} onPress={() => setShowModal(true)}>
-        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Vertical RGB</Text>
+        <Text style={{ color: '#707070', fontWeight: 'bold', textAlign: 'center' }}>Panel4</Text>
       </Pressable>
 
       <Modal onRequestClose={() => setShowModal(false)} visible={showModal} animationType='slide'>
@@ -37,39 +28,17 @@ export default function Example() {
           <View style={styles.pickerContainer}>
             <ColorPicker
               value={selectedColor.value}
-              sliderThickness={30}
-              thumbSize={30}
+              sliderThickness={25}
+              thumbSize={24}
               thumbShape='circle'
               onChange={onColorSelect}
-              thumbAnimationDuration={100}
-              adaptSpectrum
-              boundedThumb
             >
-              <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}>
-                <View style={{ alignItems: 'center' }}>
-                  <Text style={styles.sliderTitle}>R</Text>
-                  <RedSlider style={styles.sliderStyle} vertical reverse />
-                </View>
+              <Panel4 style={styles.panelStyle} thumbShape='ring' />
 
-                <View style={{ alignItems: 'center' }}>
-                  <Text style={styles.sliderTitle}>G</Text>
-                  <GreenSlider style={styles.sliderStyle} vertical reverse />
-                </View>
+              <OpacitySlider style={styles.sliderStyle} />
 
-                <View style={{ alignItems: 'center' }}>
-                  <Text style={styles.sliderTitle}>B</Text>
-                  <BlueSlider style={styles.sliderStyle} vertical reverse />
-                </View>
-
-                <View style={{ alignItems: 'center' }}>
-                  <Text style={styles.sliderTitle}>A</Text>
-                  <OpacitySlider style={styles.sliderStyle} vertical reverse />
-                </View>
-              </View>
-
-              <Swatches style={styles.swatchesContainer} swatchStyle={styles.swatchStyle} colors={customSwatches} />
               <View style={styles.previewTxtContainer}>
-                <PreviewText style={{ color: '#707070' }} colorFormat='hsva' />
+                <PreviewText style={{ color: '#707070' }} colorFormat='hwba' />
               </View>
             </ColorPicker>
           </View>
@@ -87,7 +56,6 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: 'center',
     alignContent: 'center',
-    backgroundColor: 'orange',
   },
   pickerContainer: {
     alignSelf: 'center',
@@ -105,15 +73,22 @@ const styles = StyleSheet.create({
 
     elevation: 10,
   },
-  sliderTitle: {
-    color: '#000',
-    fontWeight: 'bold',
-    marginBottom: 5,
-    paddingHorizontal: 4,
+  panelStyle: {
+    borderRadius: 16,
+
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+
+    elevation: 5,
   },
   sliderStyle: {
-    height: 300,
     borderRadius: 20,
+    marginTop: 20,
 
     shadowColor: '#000',
     shadowOffset: {
@@ -130,24 +105,6 @@ const styles = StyleSheet.create({
     marginTop: 20,
     borderTopWidth: 1,
     borderColor: '#bebdbe',
-  },
-  swatchesContainer: {
-    paddingTop: 20,
-    marginTop: 20,
-    borderTopWidth: 1,
-    borderColor: '#bebdbe',
-    alignItems: 'center',
-    flexWrap: 'nowrap',
-    gap: 10,
-  },
-  swatchStyle: {
-    borderRadius: 20,
-    height: 30,
-    width: 30,
-    margin: 0,
-    marginBottom: 0,
-    marginHorizontal: 0,
-    marginVertical: 0,
   },
   openButton: {
     width: '100%',
@@ -168,7 +125,7 @@ const styles = StyleSheet.create({
   },
   closeButton: {
     position: 'absolute',
-    bottom: 20,
+    bottom: 10,
     borderRadius: 20,
     paddingHorizontal: 40,
     paddingVertical: 10,
